Use a short[] for the audio buffer to avoid using deprecated AudioTrack functionality

+++ moonlight-common/src/main/java/com/limelight/nvstream/av/audio/AudioRenderer.java
@@ -7,7 +7,7 @@ public interface AudioRenderer {
 
 	void stop();
 	
-	void playDecodedAudio(byte[] audioData);
+	void playDecodedAudio(short[] audioData);
 	
 	void cleanup();
 }
